% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relome.R
\name{plotClinHistograms}
\alias{plotClinHistograms}
\title{plotClinHistograms}
\usage{
plotClinHistograms(data, side = 2, width = 15, height = 10,
  filename = "res.pdf", mfrow = c(4, 7), verbose = TRUE,
  col = c("blueviolet", "cornflowerblue", "black"))
}
\arguments{
\item{data:}{a data frame containing the variables to explore}

\item{side:}{2 for the columns, If something else the dataset will be transposed}

\item{width:}{pdf width}

\item{height:}{pdf height}

\item{filename:}{pdf file name}

\item{mfrow:}{pdf mfrow}

\item{verbose:}{print information on the execution process}

\item{col:}{colors for the histograms}
}
\value{
nothing
}
\description{
function that plots histograms and barplots of a given dataset for descriptive analysis.
}
\details{
plotClinHistograms function.
}

