% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relome.R
\name{resetFactors}
\alias{resetFactors}
\title{resetFactors}
\usage{
resetFactors(data, side = 2, character.2.factor = TRUE)
}
\arguments{
\item{data:}{a data frame containing the variables to explore}

\item{side:}{2 for the columns, If something else the dataset will be transposed}

\item{character.2.factor:}{should the character vectors be transformed to factor}
}
\value{
the updated data frame
}
\description{
It will recreate the factor variable to relevel
}
\details{
resetFactors function.
}
