% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relome.R
\name{runRelome}
\alias{runRelome}
\title{runRelome}
\usage{
runRelome(data, interest = "", threshold = 0.05, adjust = "BH",
  adjust.by.var = TRUE, zoom.p = TRUE, verbose = TRUE, plot = TRUE,
  mfrow = c(4, 6), width = 15, height = 10, save.all = TRUE,
  rerun.all = FALSE, col = list(scatter = "gold2", category =
  gray.colors(2)))
}
\arguments{
\item{data:}{a data frame containing the variables to explore}

\item{threshold:}{significance threshold of the variables to extract.}

\item{adjust:}{multiple testing adjustment method}

\item{adjust.by.var:}{weather to adjust only for one variable of interest or for all the tests performed.}

\item{zoom.p:}{use the p value as significance selector. If false it will be the adjusted p-value.}

\item{verbose:}{print running information.}

\item{plot:}{If TRUE produce the co-relation graphics.}

\item{mfrow:}{grid disposition of the graph (default:c(4,6))}

\item{width:}{PDF width in inches (default:15)}

\item{height:}{PDF heaight in inches (default:10)}

\item{save.all:}{weather to save intermediary files (default:TRUE)}

\item{return.all:}{weather to return all the results (default:TRUE)}

\item{col:}{a list containg colot information for the graphs, scatterplots and tables.}
}
\value{
an object containing the different results
}
\description{
Runs the different functions that are needed to compute and visualize relations beween variables.
}
\details{
main runRelome function.
}

